[refactor]將追蹤相關 API 重構為 controller 格式
